requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.1
django-import-export==3.2.0
django-js-asset==2.0.0
django-mptt==0.14.0
et-xmlfile==1.1.0
idna==3.4
loguru==0.7.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.5.0
polib==1.2.0
python-dotenv==1.0.0
PyYAML==6.0
requests==2.30.0
sqlparse==0.4.4
tablib==3.4.0
tzdata==2023.3
urllib3==2.0.2
win32-setctime==1.1.0
xlrd==2.0.1
xlwt==1.3.0

django-filter==23.2
django-property-filter==1.1.2
djangorestframework==3.14.0
setuptools==67.7.2